+++ CHANGELOG.md
@@ -11,3 +11,8 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/) and this p
 ## 3.0.2
 
 - Устранен баг - файлы json не скачивались вместе с проектом.
+
+## 3.0.3
+
+- New `Lines.getFi({ t, h })` method (Уменьшена погрешность определения относительной влажности)
+- Увеличено количество точек для `Formulas.getHumidityPoints()`